+++ internal/clients/fleet/fleet.go
@@ -66,8 +66,16 @@ func ReadAgentPolicy(ctx context.Context, client *Client, id string) (*fleetapi.
 }
 
 // CreateAgentPolicy creates a new agent policy.
-func CreateAgentPolicy(ctx context.Context, client *Client, req fleetapi.AgentPolicyCreateRequest) (*fleetapi.AgentPolicy, diag.Diagnostics) {
-	resp, err := client.API.CreateAgentPolicyWithResponse(ctx, req)
+func CreateAgentPolicy(ctx context.Context, client *Client, req fleetapi.AgentPolicyCreateRequest, sysMonitoring bool) (*fleetapi.AgentPolicy, diag.Diagnostics) {
+	resp, err := client.API.CreateAgentPolicyWithResponse(ctx, req, func(ctx context.Context, req *http.Request) error {
+		if sysMonitoring {
+			qs := req.URL.Query()
+			qs.Add("sys_monitoring", "true")
+			req.URL.RawQuery = qs.Encode()
+		}
+
+		return nil
+	})
 	if err != nil {
 		return nil, diag.FromErr(err)
 	}